consumer/requirements.txt
kafka-python==2.0.2
pandas==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
boto3==1.35.67
